+++ R/paragraph2vec.R
@@ -36,9 +36,8 @@
 #' @seealso \code{\link{predict.paragraph2vec}}, \code{\link{as.matrix.paragraph2vec}}
 #' @export
 #' @examples
-#' \dontshow{if(require(tokenizers.bpe) & require(udpipe))\{}
+#' \dontshow{if(require(tokenizers.bpe))\{}
 #' library(tokenizers.bpe)
-#' library(udpipe)
 #' ## Take data and standardise it a bit
 #' data(belgium_parliament, package = "tokenizers.bpe")
 #' str(belgium_parliament)
@@ -47,7 +46,7 @@
 #' x$text   <- gsub("[^[:alpha:]]", " ", x$text)
 #' x$text   <- gsub("[[:space:]]+", " ", x$text)
 #' x$text   <- trimws(x$text)
-#' x$nwords <- txt_count(x$text, pattern = " ")
+#' x$nwords <- txt_count_words(x$text)
 #' x <- subset(x, nwords < 1000 & nchar(text) > 0)
 #' 
 #' ## Build the model
@@ -121,12 +120,11 @@ paragraph2vec <- function(x,
 #' @seealso \code{\link{paragraph2vec}}, \code{\link{read.paragraph2vec}}
 #' @export
 #' @examples 
-#' \dontshow{if(require(tokenizers.bpe) & require(udpipe))\{}
+#' \dontshow{if(require(tokenizers.bpe))\{}
 #' library(tokenizers.bpe)
-#' library(udpipe)
 #' data(belgium_parliament, package = "tokenizers.bpe")
 #' x <- subset(belgium_parliament, language %in% "french")
-#' x <- subset(x, nchar(text) > 0 & txt_count(text, pattern = " ") < 1000)
+#' x <- subset(x, nchar(text) > 0 & txt_count_words(text) < 1000)
 #' 
 #' model <- paragraph2vec(x = x, type = "PV-DM",   dim = 15,  iter = 5)
 #' \donttest{
@@ -162,12 +160,11 @@ as.matrix.paragraph2vec_trained <- function(x, encoding='UTF-8', ...){
 #' @export
 #' @seealso \code{\link{paragraph2vec}}
 #' @examples 
-#' \dontshow{if(require(tokenizers.bpe) & require(udpipe))\{}
+#' \dontshow{if(require(tokenizers.bpe))\{}
 #' library(tokenizers.bpe)
-#' library(udpipe)
 #' data(belgium_parliament, package = "tokenizers.bpe")
 #' x <- subset(belgium_parliament, language %in% "french")
-#' x <- subset(x, nchar(text) > 0 & txt_count(text, pattern = " ") < 1000)
+#' x <- subset(x, nchar(text) > 0 & txt_count_words(text) < 1000)
 #' 
 #' \donttest{
 #' model <- paragraph2vec(x = x, type = "PV-DM",   dim = 100, iter = 20)
@@ -209,12 +206,11 @@ write.paragraph2vec <- function(x, file){
 #' }
 #' @export
 #' @examples 
-#' \dontshow{if(require(tokenizers.bpe) & require(udpipe))\{}
+#' \dontshow{if(require(tokenizers.bpe))\{}
 #' library(tokenizers.bpe)
-#' library(udpipe)
 #' data(belgium_parliament, package = "tokenizers.bpe")
 #' x <- subset(belgium_parliament, language %in% "french")
-#' x <- subset(x, nchar(text) > 0 & txt_count(text, pattern = " ") < 1000)
+#' x <- subset(x, nchar(text) > 0 & txt_count_words(text) < 1000)
 #' 
 #' \donttest{
 #' model <- paragraph2vec(x = x, type = "PV-DM",   dim = 100, iter = 20)
@@ -299,13 +295,12 @@ summary.paragraph2vec_trained <- function(object, type = "vocabulary", which = c
 #' @seealso \code{\link{paragraph2vec}}, \code{\link{read.paragraph2vec}}
 #' @export
 #' @examples 
-#' \dontshow{if(require(tokenizers.bpe) & require(udpipe))\{}
+#' \dontshow{if(require(tokenizers.bpe))\{}
 #' library(tokenizers.bpe)
-#' library(udpipe)
 #' data(belgium_parliament, package = "tokenizers.bpe")
 #' x <- belgium_parliament
 #' x <- subset(x, language %in% "dutch")
-#' x <- subset(x, nchar(text) > 0 & txt_count(text, pattern = " ") < 1000)
+#' x <- subset(x, nchar(text) > 0 & txt_count_words(text) < 1000)
 #' x$doc_id <- sprintf("doc_%s", 1:nrow(x))
 #' x$text   <- tolower(x$text)
 #' x$text   <- gsub("[^[:alpha:]]", " ", x$text)