+++ R/pkg.R
@@ -1,5 +1,5 @@
 #' @importFrom Rcpp evalCpp
-#' @importFrom stats predict
+#' @importFrom stats predict median update
 #' @useDynLib doc2vec
 NULL